+++ src/engine/ctf/mod.rs
@@ -17,6 +17,7 @@ mod challenge;
 mod event;
 mod import;
 mod resolve;
+mod shelve;
 
 use anyhow::Result;
 use chrono::prelude::*;
@@ -28,6 +29,7 @@ use std::path::Path;
 pub use archive::archive_event;
 pub use resolve::get_event_path;
 pub use challenge::{add_challenge, challenge_status, generate_writeup, solve_challenge};
+pub use shelve::shelve_challenge;
 pub use event::{
     check_active_expiry, check_expiries, create_event, find_event_root, finish_event,
     get_active_event_root, get_context_info, list_events, schedule_event, set_active_event,